Add multiple test cases for debugging two-fluid single-column RBC model
including changes to setup and plotting scripts to make more reproducible.

The cases for testing these issues are in the new directory:
 - danRun/partitioned/RayleighBenard/singleColumn/debuggingCases
The high resolution reference profiles are in directory "hiResProfiles" within that parent directory, along with a README describing where they come from. The "initHiRes" versions of test cases, and the plotting script "plot1DProfile.py", now use these profiles.

There are multiple cases in the directory; the reference case is "reference_gamma_1_e-3", which reproduces Fig. 4. Hopefully the naming conventions are clear, but for clarity:
   - "initHiRes" means "initialised from hiResProfiles"
   - "CN_[number]" means offCentre set to [number] in fvSchemes
   - "uniformSigma" means "uniformSigma" transfer used rather than div transfer
   - "effectivePrandtlNumber" means \nu = \nu_actual + \gamma, and \gamma = 0, in constant/environmentalProperties
   - "largeDiffusivities" uses a much larger thermal diffusivity as well as a larger viscosity (the latter set by \nu = \nu_actual + \gamma )